+++ breeze-core/src/main/kotlin/com/windea/breezeframework/core/extensions/AnnotationExtensions.kt
@@ -10,12 +10,12 @@ import kotlin.reflect.*
 inline fun <reified A : Annotation> KAnnotatedElement.findAnnotation(predicate: (A) -> Boolean): A? =
 	this.annotations.filterIsInstance<A>().firstOrNull(predicate)
 
-/**得到可注解元素的指定类型的注解列表。*/
+/**得到可注解元素的指定类型的注解列表。注意：Kotlin目前只能使用SOURCE级别的可重复注解。*/
 @OutlookImplementationApi
 inline fun <reified A : Annotation> KAnnotatedElement.findAnnotations(): List<A> =
 	this.annotations.filterIsInstance<A>()
 
-/**得到可注解元素的指定类型的满足指定预测的注解列表。*/
+/**得到可注解元素的指定类型的满足指定预测的注解列表。注意：Kotlin目前只能使用SOURCE级别的可重复注解。*/
 @OutlookImplementationApi
 inline fun <reified A : Annotation> KAnnotatedElement.findAnnotations(predicate: (A) -> Boolean): List<A> =
 	this.annotations.filterIsInstance<A>().filter(predicate)